HistScanner added some CFD timing.
Added branches for baseline, timestamp (timeStampNs), cfdBin, and event
time from the timestamp and cfd (timeCfdNs).

Updated HistScanner and RootScanner to handle double ranges.

Currently, the HistScanner does not properly plot timeStampNs and
timeCfdNs as they are doubles.


Former-commit-id: 6978389473dd3d629206904fe384f89468536525

+++ Scan/util/include/RootScanner.hpp
@@ -22,10 +22,10 @@ class RootScanner : public ScanInterface {
 	
 		static constexpr const int numAxes_ = 3;
 		struct AxisInfo {	
-			float rangeUserMin[numAxes_];
-			float rangeUserMax[numAxes_];
-			float limitMin[numAxes_];
-			float limitMax[numAxes_];
+			double rangeUserMin[numAxes_];
+			double rangeUserMax[numAxes_];
+			double limitMin[numAxes_];
+			double limitMax[numAxes_];
 			bool reset;
 		};
 		std::map< TVirtualPad*, AxisInfo > zoomInfo_;
